edited styles in components, created card component, edited cardlist component, edited loopline animation
