+++ src/app/page.tsx
@@ -1,22 +1,17 @@
 import Image from "next/image";
 import { Box, Container } from "@mui/material";
-import Carouselsection from "@/containers/carouselsection/Carouselsection";
+import Carouselsection from "@/containers/carouselcontainer/Carouselcontainer";
 import Herosection from "@/ui/herosection/Herosection";
-import CardList from "@/containers/listcontainer/CardList";
+import CardList from "@/containers/cardslist/CardList";
 import AboutUs from "@/ui/aboutus/AboutUs";
 import WhyWe from "@/ui/whywe/WhyWe";
-import Contacts from "@/components/contacts/Contacts";
+import Contacts from "@/ui/contacts/Contacts";
 import "./globals.css";
 import styles from "./homepage.module.css";
 import ThemeContainer from "@/containers/themecontainer/ThemeContainer";
 import waveSvg from "../../public/assets/svg/bg/waveherosec.svg";
 import tiltWHywe from "../../public/assets/svg/bg/tiltwhysec.svg";
 import LoopLine from "@/ui/loopline/LoopLine";
-
-const contrStyles = {
-  paddingTop: { xs: "40px", md: "80px" },
-  paddingBottom: { xs: "40px", md: "80px" },
-};
 
 export default function Home() {
   return (
@@ -25,8 +20,6 @@ export default function Home() {
         <section className={styles.hero_section_bg}>
           <Container
             sx={{
-              paddingTop: { xs: "40px", md: "80px" },
-              paddingBottom: { xs: "40px", md: "80px" },
               display: { xs: "block", md: "flex" },
               gap: "80px",
             }}
@@ -37,24 +30,24 @@ export default function Home() {
           </Container>
         </section>
         <section className={styles.card_list_section}>
-          <Container sx={contrStyles}>
+          <Container>
             <CardList />
           </Container>
           <LoopLine />
         </section>
         <section className={styles.aboutus_section}>
-          <Container sx={contrStyles}>
+          <Container>
             <AboutUs />
             <Image src={tiltWHywe} alt="svg" className={styles.tilt} />
           </Container>
         </section>
         <section className={styles.why_we_section}>
-          <Container sx={contrStyles}>
+          <Container>
             <WhyWe />
           </Container>
         </section>
         <section className={styles.contacts_section}>
-          <Container sx={contrStyles}>
+          <Container>
             <Contacts />
           </Container>
         </section>
