+++ src/app/services/bookingeneer/page.module.scss
@@ -1,0 +1,4 @@
+.page_container{
+    @apply my-7 md:my-14 flex items-center flex-col gap-8 md:gap-14 min-h-screen;
+}
+
