+++ src/components/card/Card.module.css
@@ -1,14 +1,3 @@
-.card_list_section {
-  position: relative;
-  z-index: 1;
-  padding: 20vh 0;
-  background: radial-gradient(50% 50% at 50% 50%, rgba(255, 255, 255, 0) 0%, #fff 100%);
-}
-
-.h4 {
-  @apply text-center my-5 md:my-10 font-bold text-gray-900;
-}
-
 .grid_container {
   @apply grid grid-cols-1 md:grid-cols-2 gap-4;
 }
@@ -54,4 +43,4 @@
 }
 .relative li {
   @apply text-base md:text-lg text-slate-500;
-}/*# sourceMappingURL=CardList.module.css.map */
+}/*# sourceMappingURL=Card.module.css.map */
