+++ src/components/carousel/item/stylesconstants.ts
