+++ src/containers/carouselcontainer/Carouselsection.module.scss
@@ -1,3 +1,5 @@
+
+
 .orange_vector {
     width: 60px;
     height: 60px;
