+++ src/containers/carouselcontainer/styleconstans.ts
