+++ src/ui/loopline/LoopLine.module.css
@@ -3,4 +3,7 @@
 }
 .loop_container p {
   @apply uppercase text-sm md:text-xl lg:text-2xl text-slate-700 font-semibold w-full whitespace-nowrap;
+}
+.loop_container span {
+  @apply text-green-500;
 }/*# sourceMappingURL=LoopLine.module.css.map */
