+++ src/ui/loopline/LoopLine.module.scss
@@ -3,4 +3,7 @@
     p {
         @apply uppercase text-sm md:text-xl lg:text-2xl text-slate-700 font-semibold w-full whitespace-nowrap;
     }
+    span{
+        @apply text-green-500;
+    }
 }
