+++ src/ui/whywe/Whywe.module.scss
@@ -2,7 +2,7 @@
     @apply text-center my-5 md:my-10 font-bold text-gray-900;
 }
 
-.contr_content{
+.contr_content {
     @apply flex flex-col m-auto gap-3 md:gap-6;
 }
 
@@ -10,8 +10,6 @@
     @apply text-center my-5 md:my-10 font-bold text-white text-center bg-slate-600 rounded-md p-2;
 }
 
-.card{
+.card {
     @apply flex text-start flex-col items-start w-80 gap-2 bg-slate-100 p-3 md:p-6 rounded-md transition-transform duration-500 hover:scale-125 cursor-pointer;
-
 }
-
